v 1.3 подключен phpstan/larastan и внесены изменения в код, удовлетворяющие 7 уровню проверки автоформатирование кода исправлена ошибка, при которой на страницах регистрации и входа не подгружались стили

+++ app/Http/Controllers/Client/API/APIUserActiveController.php
@@ -18,7 +18,7 @@ class APIUserActiveController extends Controller
     public function likedToggle(ProductType $productType): void
     {
         $this->authorize('product', User::class);
-        $this->service->likedToggle($productType, auth('api')->id());
+        $this->service->likedToggle($productType, (int)auth('api')->id());
     }
 
     public function commentStore(StoreRequest $request): void
